+++ R/fn.R
@@ -468,12 +468,8 @@ as_closure <- function(x, env = caller_env()) {
       }
 
       fmls <- formals(.ArgsEnv[[fn_name]] %||% .GenericArgsEnv[[fn_name]])
+      prim_call <- call2(fn_name, !!!syms(names(fmls)))
 
-      args <- syms(names(fmls))
-      args <- set_names(args)
-      names(args)[(names(args) == "...")] <- ""
-
-      prim_call <- call2(fn_name, splice(args))
       # The closure wrapper should inherit from the global environment
       # to ensure proper lexical dispatch with methods defined there
       new_function(fmls, prim_call, global_env())