+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # rlang 0.3.0.9000
 
+* `as_closure()` wrappers now call primitives with positional
+  arguments to avoid edge case issues of argument matching.
+
 * `as_closure()` wrappers now dispatch properly on methods defined in the global environment (tidyverse/purrr#459).
 
 * `as_closure()` now supports both base-style (`e1` and `e2`) and